pyproject.toml
[project]
name = "andrewcouncil-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "click>=8.2.1",
  "python-frontmatter>=1.1.0",
  "toml>=0.10.2"
]
